Update scripts to install nvim from source
Some of the plugins and features of neovim that I user
are not available in the neovim versions in Ubuntu's
repositories. I could add neovim unstable in order to
install from apt, but I find it nice to be able to have
complete control over the version that I have installed
on my system through git.