Update telescope keybindings
Ctrl+o is no longer used to open a file as I learned that
Ctrl+o+o can be used to open the last file to the last location
you were editing last time neovim was open. I found the Ctrl+o
keybinding in telescope to interfere with this functionality.